Module aws_sdk_route53recoveryreadiness::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

User does not have sufficient access to perform this action.

ConflictException

Updating or deleting a resource can cause an inconsistent state.

CreateCellError

CreateCrossAccountAuthorizationError

CreateReadinessCheckError

CreateRecoveryGroupError

CreateResourceSetError

DeleteCellError

DeleteCrossAccountAuthorizationError

DeleteReadinessCheckError

DeleteRecoveryGroupError

DeleteResourceSetError

GetArchitectureRecommendationsError

GetCellError

GetCellReadinessSummaryError

GetReadinessCheckError

GetReadinessCheckResourceStatusError

GetReadinessCheckStatusError

GetRecoveryGroupError

GetRecoveryGroupReadinessSummaryError

GetResourceSetError

InternalServerException

An unexpected error occurred.

ListCellsError

ListCrossAccountAuthorizationsError

ListReadinessChecksError

ListRecoveryGroupsError

ListResourceSetsError

ListRulesError

ListTagsForResourcesError

ResourceNotFoundException

The requested resource does not exist.

TagResourceError

ThrottlingException

Request was denied due to request throttling.

UntagResourceError

UpdateCellError

UpdateReadinessCheckError

UpdateRecoveryGroupError

UpdateResourceSetError

ValidationException

The input fails to satisfy the constraints specified by an AWS service.

Enums

CreateCellErrorKind

CreateCrossAccountAuthorizationErrorKind

CreateReadinessCheckErrorKind

CreateRecoveryGroupErrorKind

CreateResourceSetErrorKind

DeleteCellErrorKind

DeleteCrossAccountAuthorizationErrorKind

DeleteReadinessCheckErrorKind

DeleteRecoveryGroupErrorKind

DeleteResourceSetErrorKind

GetArchitectureRecommendationsErrorKind

GetCellErrorKind

GetCellReadinessSummaryErrorKind

GetReadinessCheckErrorKind

GetReadinessCheckResourceStatusErrorKind

GetReadinessCheckStatusErrorKind

GetRecoveryGroupErrorKind

GetRecoveryGroupReadinessSummaryErrorKind

GetResourceSetErrorKind

ListCellsErrorKind

ListCrossAccountAuthorizationsErrorKind

ListReadinessChecksErrorKind

ListRecoveryGroupsErrorKind

ListResourceSetsErrorKind

ListRulesErrorKind

ListTagsForResourcesErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateCellErrorKind

UpdateReadinessCheckErrorKind

UpdateRecoveryGroupErrorKind

UpdateResourceSetErrorKind