Module aws_sdk_route53recoveryreadiness::model

Modules

cell_output

See CellOutput

dns_target_resource

See DnsTargetResource

list_rules_output

See ListRulesOutput

message

See Message

nlb_resource

See NlbResource

r53_resource_record

See R53ResourceRecord

readiness_check_output

See ReadinessCheckOutput

readiness_check_summary

See ReadinessCheckSummary

recommendation

See Recommendation

recovery_group_output

See RecoveryGroupOutput

resource

See Resource

resource_result

See ResourceResult

resource_set_output

See ResourceSetOutput

rule_result

See RuleResult

target_resource

See TargetResource

Structs

CellOutput

A Cell and its properties

DnsTargetResource

A component for DNS/Routing Control Readiness Checks

ListRulesOutput

A collection of rules used in a readiness check

Message

Information relating to readiness check status

NlbResource

The NLB resource a DNS Target Resource points to

R53ResourceRecord

The Route 53 resource a DNS Target Resource record points to

ReadinessCheckOutput

A resource used for checking the readiness of a Resource Set

ReadinessCheckSummary

Summary of ReadinessCheck status, paginated in GetRecoveryGroupReadinessSummary and GetCellReadinessSummary

Recommendation

Guidance for improving Recovery Group resilliancy

RecoveryGroupOutput

A Recovery Group generally containing multiple Cells

Resource

The resource element of a ResourceSet

ResourceResult

Result with status for an individual resource.

ResourceSetOutput

A collection of resources of the same type

RuleResult

Result with status for an individual rule..

TargetResource

The target resource the R53 record points to

Enums

Readiness

The readiness of an entire ReadinessCheck or an individual resource ARN.