Module aws_sdk_route53recoveryreadiness::operation

Structs

CreateCell

Creates a new Cell.

CreateCrossAccountAuthorization

Create a new cross account readiness authorization.

CreateReadinessCheck

Creates a new Readiness Check.

CreateRecoveryGroup

Creates a new Recovery Group.

CreateResourceSet

Creates a new Resource Set.

DeleteCell

Deletes an existing Cell.

DeleteCrossAccountAuthorization

Delete cross account readiness authorization

DeleteReadinessCheck

Deletes an existing Readiness Check.

DeleteRecoveryGroup

Deletes an existing Recovery Group.

DeleteResourceSet

Deletes an existing Resource Set.

GetArchitectureRecommendations

Returns a collection of recommendations to improve resilliance and readiness check quality for a Recovery Group.

GetCell

Returns information about a Cell.

GetCellReadinessSummary

Returns information about readiness of a Cell.

GetReadinessCheck

Returns information about a ReadinessCheck.

GetReadinessCheckResourceStatus

Returns detailed information about the status of an individual resource within a Readiness Check’s Resource Set.

GetReadinessCheckStatus

Returns information about the status of a Readiness Check.

GetRecoveryGroup

Returns information about a Recovery Group.

GetRecoveryGroupReadinessSummary

Returns information about a Recovery Group.

GetResourceSet

Returns information about a Resource Set.

ListCells

Returns a collection of Cells.

ListCrossAccountAuthorizations

Returns a collection of cross account readiness authorizations.

ListReadinessChecks

Returns a collection of Readiness Checks.

ListRecoveryGroups

Returns a collection of Recovery Groups.

ListResourceSets

Returns a collection of Resource Sets.

ListRules

Returns a collection of rules that are applied as part of Readiness Checks.

ListTagsForResources

Returns a list of the tags assigned to the specified resource.

TagResource

Adds tags to the specified resource. You can specify one or more tags to add.

UntagResource

Removes tags from the specified resource. You can specify one or more tags to remove.

UpdateCell

Updates an existing Cell.

UpdateReadinessCheck

Updates an exisiting Readiness Check.

UpdateRecoveryGroup

Updates an existing Recovery Group.

UpdateResourceSet

Updates an existing Resource Set.